# vtn/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: OpenADR 3 API
  description: |
    The OpenADR 3 API supports energy retailer to energy customer Demand Response programs.
    See OpenADR 3 User Guide and Defintions for detailed descriptions of usage.
    The API includes the following capabilities and operations:

    __Manage programs:__

    * Create/Update/Delete a program
    * Search programs

    __Manage events:__

    * Create/Update/Delete an event
    * Search events

    __Manage reports:__

    * Create/Update/Delete a report
    * Search reports

    __Manage subscriptions:__

    * Create/Update/Delete subscriptions to programs, events, and reports
    * Search subscriptions
    * Subscriptions allows clients to register a callback URL (webhook) to be notified
      on the change of state of a resource

    __Manage vens:__

    * Create/Update/Delete vens and ven resources
    * Search ven and ven resources

    __Manage tokens:__

    * Obtain an access token
    * This endpoint is provided as a convenience and may be neglected in a commercial implementation
  contact:
    email: frank@pajaritotech.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 3.0.1
servers:
- url: http://localhost:8081/openadr3/3.0.1
  description: base path
paths:
  /programs:
    get:
      tags:
      - programs
      summary: searches all programs
      description: |
        List all programs known to the server.
        May filter results by targetType and targetValues as query params.
        Use skip and pagination query params to limit response size.
      operationId: search_all_programs
      parameters:
      - name: targetType
        in: query
        description: "Indicates targeting type, e.g. GROUP"
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: targetValues
        in: query
        description: "List of target values, e.g. group names"
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      - name: skip
        in: query
        description: number of records to skip for pagination.
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: limit
        in: query
        description: maximum number of records to return.
        required: false
        style: form
        explode: true
        schema:
          maximum: 50
          minimum: 0
          type: integer
          format: int32
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/program'
                x-content-type: application/json
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - read_all
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.programs_controller
    post:
      tags:
      - programs
      summary: create a program
      description: Create a new program in the server.
      operationId: create_program
      requestBody:
        description: program item to add.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/program'
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/program'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "409":
          description: Conflict. Implementation dependent response if program with
            the same programName exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - write_programs
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.programs_controller
  /programs/{programID}:
    get:
      tags:
      - programs
      summary: searches programs by program ID
      description: |
        Fetch the program specified by the programID in path.
      operationId: search_program_by_program_id
      parameters:
      - name: programID
        in: path
        description: Object ID of the program object.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/program'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - read_all
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.programs_controller
    put:
      tags:
      - programs
      summary: update a program
      description: Update an existing program with the programID in path.
      operationId: update_program
      parameters:
      - name: programID
        in: path
        description: Object ID of the program object.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      requestBody:
        description: program item to update.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/program'
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/program'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "409":
          description: Conflict. Implementation dependent response if program with
            the same programName exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - write_programs
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.programs_controller
    delete:
      tags:
      - programs
      summary: delete a program
      description: Delete an existing program with the programID in path.
      operationId: delete_program
      parameters:
      - name: programID
        in: path
        description: Object ID of the program object.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/program'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - write_programs
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.programs_controller
...
components:
  schemas:
    program:
      required:
      - programName
      type: object
      properties:
        id:
          $ref: '#/components/schemas/objectID'
        createdDateTime:
          $ref: '#/components/schemas/dateTime'
        modificationDateTime:
          $ref: '#/components/schemas/dateTime'
        objectType:
          type: string
          description: Used as discriminator
          enum:
          - PROGRAM
        programName:
          maxLength: 128
          minLength: 1
          type: string
          description: Short name to uniquely identify program.
          example: ResTOU
        programLongName:
          type: string
          description: Long name of program for human readability.
          nullable: true
          example: Residential Time of Use-A
        retailerName:
          type: string
          description: Short name of energy retailer providing the program.
          nullable: true
          example: ACME
        retailerLongName:
          type: string
          description: Long name of energy retailer for human readability.
          nullable: true
          example: ACME Electric Inc.
        programType:
          type: string
          description: A program defined categorization.
          nullable: true
          example: PRICING_TARIFF
        country:
          type: string
          description: Alpha-2 code per ISO 3166-1.
          nullable: true
          example: US
        principalSubdivision:
          type: string
          description: Coding per ISO 3166-2. E.g. state in US.
          nullable: true
          example: CO
        timeZoneOffset:
          $ref: '#/components/schemas/duration'
        intervalPeriod:
          $ref: '#/components/schemas/intervalPeriod'
        programDescriptions:
          type: array
          description: A list of programDescriptions
          nullable: true
          items:
            $ref: '#/components/schemas/program_programDescriptions'
        bindingEvents:
          type: boolean
          description: True if events are fixed once transmitted.
          nullable: true
          example: false
        localPrice:
          type: boolean
          description: True if events have been adapted from a grid event.
          nullable: true
          example: false
        payloadDescriptors:
          type: array
          description: A list of payloadDescriptors.
          nullable: true
          items:
            discriminator:
              propertyName: objectType
            anyOf:
            - $ref: '#/components/schemas/eventPayloadDescriptor'
            - $ref: '#/components/schemas/reportPayloadDescriptor'
        targets:
          type: array
          description: A list of valuesMap objects.
          nullable: true
          items:
            $ref: '#/components/schemas/valuesMap'
      description: Provides program specific metadata from VTN to VEN.
      example:
        country: US
        programType: PRICING_TARIFF
        bindingEvents: false
        programLongName: Residential Time of Use-A
        timeZoneOffset: PT1H
        createdDateTime: 2023-06-15T09:30:00Z
        modificationDateTime: null
        targets:
        - values:
          - 0.17
          type: PRICE
        - values:
          - 0.17
          type: PRICE
        objectType: PROGRAM
        payloadDescriptors:
        - ""
        - ""
        principalSubdivision: CO
        programName: ResTOU
        retailerName: ACME
        id: object-999
        programDescriptions:
        - URL: www.myCorporation.com/myProgramDescription
        - URL: www.myCorporation.com/myProgramDescription
        intervalPeriod:
          duration: null
          randomizeStart: null
          start: null
        localPrice: false
        retailerLongName: ACME Electric Inc.
...
    intervalPeriod:
      required:
      - start
      type: object
      properties:
        start:
          $ref: '#/components/schemas/dateTime'
        duration:
          $ref: '#/components/schemas/duration'
        randomizeStart:
          $ref: '#/components/schemas/duration'
      description: |
        Defines temporal aspects of intervals.
        A duration of default PT0S indicates instantaneous or infinity, depending on payloadType.
        A randomizeStart of default null indicates no randomization.
      example:
        duration: null
        randomizeStart: null
        start: null
    valuesMap:
      required:
      - type
      - values
      type: object
      properties:
        type:
          maxLength: 128
          minLength: 1
          type: string
          description: |
            Enumerated or private string signifying the nature of values.
            E.G. "PRICE" indicates value is to be interpreted as a currency.
          example: PRICE
        values:
          type: array
          description: A list of data points. Most often a singular value such as
            a price.
          example:
          - 0.17
          items:
            anyOf:
            - type: number
            - type: integer
            - type: string
            - type: boolean
            - $ref: '#/components/schemas/point'
      description: |
        Represents one or more values associated with a type.
        E.g. a type of PRICE contains a single float value.
      example:
        values:
        - 0.17
        type: PRICE
    point:
      required:
      - x
      - "y"
      type: object
      properties:
        x:
          type: number
          description: A value on an x axis.
          format: float
          example: 1.0
        "y":
          type: number
          description: A value on a y axis.
          format: float
          example: 2.0
      description: A pair of floats typically used as a point on a 2 dimensional grid.
    eventPayloadDescriptor:
      required:
      - payloadType
      type: object
      properties:
        objectType:
          type: string
          description: Used as discriminator.
          enum:
          - EVENT_PAYLOAD_DESCRIPTOR
        payloadType:
          maxLength: 128
          minLength: 1
          type: string
          description: Enumerated or private string signifying the nature of values.
          example: PRICE
        units:
          type: string
          description: Units of measure.
          nullable: true
          example: KWH
        currency:
          type: string
          description: Currency of price payload.
          nullable: true
          example: USD
      description: |
        Contextual information used to interpret event valuesMap values.
        E.g. a PRICE payload simply contains a price value, an
        associated descriptor provides necessary context such as units and currency.
      example:
        payloadType: PRICE
        currency: USD
        units: KWH
        objectType: EVENT_PAYLOAD_DESCRIPTOR
    reportPayloadDescriptor:
      required:
      - payloadType
      type: object
      properties:
        objectType:
          type: string
          description: Used as discriminator.
          enum:
          - REPORT_PAYLOAD_DESCRIPTOR
        payloadType:
          maxLength: 128
          minLength: 1
          type: string
          description: Enumerated or private string signifying the nature of values.
          example: USAGE
        readingType:
          type: string
          description: Enumerated or private string signifying the type of reading.
          nullable: true
          example: DIRECT_READ
        units:
          type: string
          description: Units of measure.
          nullable: true
          example: KWH
        accuracy:
          type: number
          description: A quantification of the accuracy of a set of payload values.
          format: float
          nullable: true
          example: 0.0
        confidence:
          maximum: 100
          minimum: 0
          type: integer
          description: A quantification of the confidence in a set of payload values.
          format: int32
          nullable: true
          example: 100
      description: |
        Contextual information used to interpret report payload values.
        E.g. a USAGE payload simply contains a usage value, an
        associated descriptor provides necessary context such as units and data quality.
      example:
        payloadType: USAGE
        confidence: 100
        accuracy: 0.0
        units: KWH
        readingType: DIRECT_READ
        objectType: REPORT_PAYLOAD_DESCRIPTOR
...
    objectID:
      maxLength: 128
      minLength: 1
      pattern: "^[a-zA-Z0-9_-]*$"
      type: string
      description: URL safe VTN assigned object ID.
      example: object-999
...
    dateTime:
      type: string
      description: datetime in ISO 8601 format
      format: date-time
      example: 2023-06-15T09:30:00Z
    duration:
      pattern: "^(-?)P(?=\\d|T\\d)(?:(\\d+)Y)?(?:(\\d+)M)?(?:(\\d+)([DW]))?(?:T(?:(\\\
        d+)H)?(?:(\\d+)M)?(?:(\\d+(?:\\.\\d+)?)S)?)?$"
      type: string
      description: duration in ISO 8601 format
      example: PT1H
      default: PT0S
...
    problem:
      type: object
      properties:
        type:
          type: string
          description: |
            An absolute URI that identifies the problem type.
            When dereferenced, it SHOULD provide human-readable documentation for the problem type
            (e.g., using HTML).
          format: uri
          example: https://zalando.github.io/problem/constraint-violation
          default: about:blank
        title:
          type: string
          description: |
            A short, summary of the problem type. Written in english and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized); example: Service Unavailable.
        status:
          maximum: 600
          exclusiveMaximum: true
          minimum: 100
          type: integer
          description: |
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          format: int32
          example: 503
        detail:
          type: string
          description: |
            A human readable explanation specific to this occurrence of the
            problem.
          example: Connection to database timed out
        instance:
          type: string
          description: |
            An absolute URI that identifies the specific occurrence of the problem.
            It may or may not yield further information if dereferenced.
          format: uri
      description: |
        reusable error response. From https://opensource.zalando.com/problem/schema.yaml.
    program_programDescriptions:
      required:
      - URL
      properties:
        URL:
          type: string
          description: A human or machine readable program description
          format: uri
          example: www.myCorporation.com/myProgramDescription
      example:
        URL: www.myCorporation.com/myProgramDescription
...
  securitySchemes:
    oAuth2ClientCredentials:
      type: oauth2
      description: Client credential flow.
      flows:
        clientCredentials:
          tokenUrl: auth/token
          scopes:
            read_all: VENs and BL can read all resources
            write_programs: Only BL can write to programs
            write_events: Only BL can write to events
            write_reports: only VENs can write to reports
            write_subscriptions: VENs and BL can write to subscriptions
            write_vens: VENS and BL can write to vens and resources
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_oAuth2ClientCredentials
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_oAuth2ClientCredentials
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_bearerAuth